* commit '57229ff7974212aff1a79e9e378edb8e83884418':
  Revert "msm: camera: isp: Disable CSI2 RX in CSID stop path"
  msm: camera: ope: Prevent OOB read in command buffer
  msm: camera: icp: Skip sending HFI commands if FW is re-downloaded
  msm: camera: isp: Disable CSI2 RX in CSID stop path
  msm: camera: isp: Handle EOF IRQ enable command as no-op on v1
  msm: camera: isp: fix null deref in IFE out resource acquire
  msm: camera: isp: Fix OOB read in dual config command buffer validation
  msm: camera: icp: Reset AXI floor level on ICP device clock disable
  msm: camera: ife: Validate BW_CONFIG blob length before parsing
  msm: camera: isp: Fix lock imbalance and NULL dereference issues
  msm: camera: jpeg: Add Null check to prevent kernel panic
  msm: camera: isp: Fix session dump buffer size checker
  msm: camera: flash: Fix issues for kthread used by precise flash
  msm: camera: isp: Fix lock imbalance and NULL dereference issues
  msm: camera: isp: Record unserved RUP in substate switching window

+++ qcom/opensource/camera-kernel/drivers/cam_req_mgr/cam_req_mgr_core.c
@@ -6272,7 +6272,7 @@ static int cam_req_mgr_dump_state_monitor_info(
 	struct cam_common_hw_dump_args  dump_args;
 	size_t                          buf_len;
 	size_t                          remain_len;
-	uint32_t                        min_len;
+	size_t                          min_len;
 	uintptr_t                       cpu_addr;
 
 	rc = cam_mem_get_cpu_buf(dump_info->buf_handle,
@@ -6296,7 +6296,7 @@ static int cam_req_mgr_dump_state_monitor_info(
 		sizeof(struct cam_common_hw_dump_header) * CAM_CRM_DUMP_EVENT_NUM_HEADERS));
 
 	if (remain_len < min_len) {
-		CAM_WARN(CAM_CRM, "Dump buffer exhaust remain %zu min %u",
+		CAM_WARN(CAM_CRM, "Dump buffer exhaust remain %zu min %zu",
 			remain_len, min_len);
 		cam_mem_put_cpu_buf(dump_info->buf_handle);
 		return -ENOSPC;
